requirements.txt


boto3==1.34.89
botocore==1.34.89
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.31.1
